♻️ merge modules `show` and `nodes`
